ongoing schedule updates
