+++ schedulite.clientapp/src/types/schedule.type.ts
@@ -4,5 +4,6 @@ export default interface ISchedule {
     activeCourses: ICourse[]
     tentativeCourses: ICourse[]
     scheduleName: string;
+    year: string;
     semester: string;
 }
